+++ src/kernel/idt.h
@@ -28,6 +28,6 @@ void invalid_opcode_handler_main(unsigned int esp);
 void general_protection_handler_main(unsigned int esp);
 void page_fault_handler_main(unsigned int esp);
 void irq0_handler_main(unsigned int esp);
-void irq1_handler_main(void);
+void irq1_handler_main(unsigned int esp);
 
 #endif /* IDT_H */